Separate primitive variants from component
