+++ packages/profile/src/components/provider/index.tsx
@@ -49,4 +49,3 @@ export function Provider({ children }: PropsWithChildren) {
     </PostHogProvider>
   );
 }
-
