+++ packages/ui-next/src/components/primitives/alert-dialog.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
 
 import { cn } from "@/utils";
-import { buttonVariants } from "@/components/primitives/button";
+import { buttonVariants } from "@/components/primitives/button/utils";
 
 const AlertDialog = AlertDialogPrimitive.Root;
